fix dru sha dk pal warr mag talent tree format, fix starter talents, adjust tank sha ui images

+++ livescripts/dh-cachedata/dh-cache.ts
@@ -260,7 +260,7 @@ export class DHCache {
     }
 
     public IsStarter(Player: TSPlayer, Talent: DHTalent) : bool {
-        let Spec : uint32 = Player.GetUInt(`Spec`)
+        let Spec : uint64 = Player.GetUInt(`Spec`)
         return Talent.Starter.includes(Spec)
     }
